luxe-digital/requirements.txt
requests==2.25.1
Scrapy==2.5.1
azure-servicebus==7.3.4